Remove `Component` trait from modules below `world` level
The `Component` trait only exists so that the world API can be more type
directed, so that the only thing you need to know to check out a component
storage is the component type.

Below the world level, there is no need to even mention "components", storages
are simply map-like storages for low-valued u32 indexes that can be joined and
optionally tracked.