`Source` associated type on `FetchResources` makes more sense as a generic parameter
This matches the `System` trait now in the `par_seq` module, and generally
should work more easily with source types with internal lifetimes.  Also it
simply makes more sense because it makes sense for a single resource to be
fetched from multiple sources (but for each source it can be fetched from, it
should describe *what* it fetches in a predictable way, i.e. the `Resources`
associated type.)