Add dynamic component sets that are self-inserting
Allows you to have "prefab" component sets or component sets that can be
manipulated without knowing the types of every component and without having to
have mutable access to a World.

Essential to implement a deferred-insert scheme, where systems may create entities
with any valid components which are then added at the end of the update.